fix cloud sync
