+++ src/popup/delegate/localStorageDelegate.js
@@ -1,16 +1,6 @@
 import { StorageDelegate } from "./storageDelegate";
 
-class LocalStorageDelegate extends StorageDelegate {
-    constructor(){
-        this.get = getLocalStorageData;
-        this.set = setLocalStorageData;
-    }
-}
-
-const localStorageDelegate = new LocalStorageDelegate();
-export default localStorageDelegate;
-
-const getLocalStorageData = async (key) => {
+export const getLocalStorageData = async (key) => {
     return new Promise((resolve, reject) => {
         chrome.storage.local.get(key, (result) => {
             if (result === undefined || result[key] === undefined) {
@@ -24,9 +14,20 @@ const getLocalStorageData = async (key) => {
     });
 }
 
-const setLocalStorageData = async (key, val) => {
+export const setLocalStorageData = async (key, val) => {
     return new Promise(() => {
         chrome.storage.local.set({ [key]: val });
         resolve();
     }).catch(e => console.log(e));
 }
+
+class LocalStorageDelegate extends StorageDelegate {
+    constructor(){
+        super();
+        this.get = getLocalStorageData;
+        this.set = setLocalStorageData;
+    }
+}
+
+const localStorageDelegate = new LocalStorageDelegate();
+export default localStorageDelegate;
