+++ src/popup/delegate/storageDelegate.js
@@ -1,6 +1,3 @@
-import cloudStorageDelegate from "./cloudStorageDelegate.js";
-import localStorageDelegate from "./localStorageDelegate";
-
 export class StorageDelegate {
     constructor(){
         this.get = async (key) => null;
@@ -8,14 +5,3 @@ export class StorageDelegate {
     }
 }
 
-export const syncStorage = async (sd1, sd2, key, merger) => {
-    const data1 = await sd1.get(key);
-    const data2 = await sd2.get(key);
-    const merged = merger(data1, data2);
-    await sd1.set(key, merged);
-    await sd2.set(key, merged);
-}
-
-export const syncLocalAndCloudStorage = async (key, merger) => {
-    await syncStorage(localStorageDelegate, cloudStorageDelegate, key, merger);
-}
